seperate code and make it more readable
