+++ components/carousel/src/main/java/com/guru/composecookbook/carousel/SelectionState.kt
@@ -1,0 +1,6 @@
+package com.guru.composecookbook.carousel
+
+enum class SelectionState {
+    Selected,
+    Undecided
+}
